Fixed quick search performance issue
Quick search w/ multiple attribute types had a performance issue
because of the way the SQL query was written. The /* ordered */ hint,
combined with the order in which tables were actually given in the
FROM clause, was forcing cartesianing. This FROM clause has been
reordered for optimal efficiency. (The query in question went from
taking 3.1 hours to taking 0.3 seconds.)