bug[ats_B0N58]: Fix tree view in artifact search
Items that appeared in list view would not be present in tree view if they
had multiple parents.  Added a new fake artifact parent to contain such
items in the tree view.  Also made orphans be detected so that they will
be put in a similar fake parent.

Found and fixed a minor bug in Artifact's isOrphan method in which it was
reporting the default root as an orphan.

Change-Id: If5591119e88285d3f1bfa24d1daeb57a8be6bab9