feature[ats_97XC2]: Fix bundle loading issues
There are several fixes here.  The first is a performance issue because we
were cycling through perspectives more than necessary to close views. The
second is we cleaned up an issue with progress monitors that caused the
status to not get properly removed.  The third was to clean up the preClose
listener to not use a job because that caused some strange shutdown behavior
because of the thread behavior that caused errors deep in the eclipse
workbench code.

Change-Id: I8c503adabf9dad7487da398f5635d7f72b04047b