bug[ats_M9QX7]: Arbitration response hangs
Change-Id: I9304ce0cc273028601eb306a549a5508e76ef2b3

When a client requests arbitration, the server checks to see what servers
are alive.  In some cases, other app servers may be slow to respond but
a socket connection has been established.  This change sets a socket read
timeout of 1 second.  If the socket connection cannot be established, the
GetMethod will fail as expected.