feature[ats_SF2V6]: Use OrcsTypes in OrcsDataStore
OrcsDataStore needs OrcsTypes which are created in the orcs.core bundle.
The orcs.core bundle needs the OrcsDataStore in order to create the
OrcsTypes.  Once the data store is up, orcs api can come up, create the
orcs types and pass them down to the data store.  The datastore needs
these types in order to set the data proxy on the attributes.

Change-Id: Ie208abba6c4c73d81ebbd3b2e07e6fe6dd4cbb85