bug[ats_ATS45072]: ATS branch commit should not overwrite conflicts
When ATS branch commit is called with the no popup option, branch
commit is called with the overwrite conflicts option set to true.
Overwritting of conflicts is an admin only capability which should
be clearly advertised by the API. Users should be notified
when they are about to run an operation that might overwrite changes.
The code was changed to never allow conflict overwrites.

Change-Id: I80d01d5edacfb96438ab1272c2e9b785b7b5b412